Adapt to work w/ suggest2
This commit adapts the search() function to use LC's suggest2 service, rather than a combination of the older suggest and didyoumean services.

The commit also 1) removes the option to search both names and subjects, because suggest2 does not accept that option, and 2) takes extra pains to avoid returning subdivisions with the same names as main headings.

To do: Rewrite documentation
To do: Test more extensively